requirements.txt
torch>=2.0.0
torchvision>=0.15.0
einops>=0.7.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
scipy>=1.11.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0
datasets>=2.14.0
huggingface-hub>=0.16.0
av>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
accelerate>=0.21.0
